Fixing signabot

+++ publish.py
@@ -1,10 +1,10 @@
 import shutil
 import os
-import sys
+import hashlib
 from typing import List
 
-path:List[str] = ["", "scripts", "python", "news_crawler"]
-#path:List[str] = ["D:", "test", "copytree"]
+#path:List[str] = ["", "scripts", "python", "news_crawler"]
+path:List[str] = ["D:", "test", "copytree"]
 baseFolder = os.sep.join(path)
 
 currentFolder:str = os.path.abspath(os.curdir)
@@ -22,6 +22,14 @@ def get_files(path:str, folder:str='.', ignoreFolders:List[str] = None):
             for r in get_files(dir, os.sep.join([folder, f]), ignoreFolders):
                 yield r
 
+def get_file_hash(filename:str) -> str:
+    with open(filename, 'rb') as f:
+        buffer = f.read()
+        return hashlib.md5(buffer).hexdigest()
+
+def are_distincts(src:str, dest:str) -> bool:
+    return get_file_hash(src) != get_file_hash(dest)
+
 def copytree(src:str, dest:str):
     for i in get_files(src, ignoreFolders=['.git', '__pycache__']):
         baseFolder, folder, file = i
@@ -32,6 +40,9 @@ def copytree(src:str, dest:str):
         sourceFile = os.sep.join([baseFolder, file])
         destinationFile = os.sep.join([destinationFolder, file])
         if os.path.isfile(destinationFile):
+            if not are_distincts(sourceFile, destinationFile):
+                print(f'Ignoring {destinationFile} cause no diff')
+                continue
             os.unlink(destinationFile)
         shutil.copy(sourceFile, destinationFile)
